Move csv-specific classes to csv package
